+++ src/queries/getFromAST.ts
@@ -2,8 +2,13 @@ import {
   DocumentNode,
   OperationDefinitionNode,
   FragmentDefinitionNode,
+  ValueNode,
 } from 'graphql';
 
+
+import {
+  valueToObjectRepresentation,
+} from '../data/storeUtils';
 
 export function getMutationDefinition(doc: DocumentNode): OperationDefinitionNode {
   checkDocument(doc);
@@ -228,4 +233,25 @@ export function getFragmentQueryDocument(document: DocumentNode, fragmentName?: 
   };
 
   return query;
+}
+
+export function getDefaultValues(definition: OperationDefinitionNode): { [key: string]: any } {
+  if (definition.variableDefinitions && definition.variableDefinitions.length) {
+    const defaultValues = definition.variableDefinitions
+      .filter(({ defaultValue }) => defaultValue)
+      .map(({ variable, defaultValue }) : { [key: string]: any } => {
+        const defaultValueObj: { [key: string]: any } = {};
+        valueToObjectRepresentation(
+          defaultValueObj,
+          variable.name,
+          defaultValue as ValueNode,
+        );
+
+        return defaultValueObj;
+      });
+
+    return Object.assign({}, ...defaultValues);
+  }
+
+  return {};
 }