+++ test/readFromStore.ts
@@ -111,6 +111,47 @@ describe('reading from the store', () => {
     });
   });
 
+  it('runs a basic query with default values for arguments', () => {
+    const query = gql`
+      query someBigQuery(
+        $stringArg: String = "This is a default string!",
+        $intArg: Int = 0,
+        $floatArg: Float,
+      ){
+        id,
+        stringField(arg: $stringArg),
+        numberField(intArg: $intArg, floatArg: $floatArg),
+        nullField
+      }
+    `;
+
+    const variables = {
+      floatArg: 3.14,
+    };
+
+    const store = {
+      'ROOT_QUERY': {
+        id: 'abcd',
+        nullField: null,
+        'numberField({"intArg":0,"floatArg":3.14})': 5,
+        'stringField({"arg":"This is a default string!"})': 'Heyo',
+      },
+    } as NormalizedCache;
+
+    const result = readQueryFromStore({
+      store,
+      query,
+      variables,
+    });
+
+    assert.deepEqual(result, {
+      id: 'abcd',
+      nullField: null,
+      numberField: 5,
+      stringField: 'Heyo',
+    });
+  });
+
   it('runs a nested query', () => {
     const result: any = {
       id: 'abcd',